[java-shell] Move JS <-> Java conversions to MongoShellConverter
