+++ packages/java-shell/src/main/kotlin/com/mongodb/mongosh/service/BaseMongoIterableHelper.kt
@@ -1,22 +1,24 @@
 package com.mongodb.mongosh.service
 
-import com.mongodb.*
+import com.mongodb.CursorType
+import com.mongodb.ReadPreference
+import com.mongodb.TagSet
 import com.mongodb.client.AggregateIterable
 import com.mongodb.client.FindIterable
 import com.mongodb.client.MongoDatabase
 import com.mongodb.client.MongoIterable
-import com.mongodb.mongosh.MongoShellEvaluator
+import com.mongodb.mongosh.MongoShellConverter
 import org.bson.Document
 import org.graalvm.polyglot.Value
 
-internal abstract class BaseMongoIterableHelper<T : MongoIterable<*>>(val iterable: T, protected val evaluator: MongoShellEvaluator, protected val options: Document?) {
+internal abstract class BaseMongoIterableHelper<T : MongoIterable<*>>(val iterable: T, protected val converter: MongoShellConverter, protected val options: Document?) {
     abstract val converters: Map<String, (T, Any?) -> Either<T>>
     abstract val defaultConverter: (T, String, Any?) -> Either<T>
 
     fun map(function: Value): BaseMongoIterableHelper<*> {
         return helper(iterable.map { v ->
-            evaluator.extract(function.execute(evaluator.toJs(v))).value
-        }, evaluator)
+            converter.toJava(function.execute(converter.toJs(v))).value
+        }, converter)
     }
 
     fun itcount(): Int {
@@ -53,8 +55,8 @@ internal abstract class BaseMongoIterableHelper<T : MongoIterable<*>>(val iterab
 }
 
 internal class MongoIterableHelper(iterable: MongoIterable<*>,
-                                   evaluator: MongoShellEvaluator,
-                                   options: Document?) : BaseMongoIterableHelper<MongoIterable<*>>(iterable, evaluator, options) {
+                                   converter: MongoShellConverter,
+                                   options: Document?) : BaseMongoIterableHelper<MongoIterable<*>>(iterable, converter, options) {
     override val converters = iterableConverters
     override val defaultConverter = unrecognizedField<MongoIterable<*>>("iterable options")
 }
@@ -64,10 +66,10 @@ internal data class AggregateCreateOptions(val db: MongoDatabase,
                                            val pipeline: List<Document>)
 
 internal class AggregateIterableHelper(iterable: AggregateIterable<*>,
-                                       evaluator: MongoShellEvaluator,
+                                       converter: MongoShellConverter,
                                        options: Document?,
                                        private val createOptions: AggregateCreateOptions?)
-    : BaseMongoIterableHelper<AggregateIterable<out Any?>>(iterable, evaluator, options) {
+    : BaseMongoIterableHelper<AggregateIterable<out Any?>>(iterable, converter, options) {
     override val converters = aggregateConverters
     override val defaultConverter = aggregateDefaultConverter
 
@@ -89,7 +91,7 @@ internal class AggregateIterableHelper(iterable: AggregateIterable<*>,
         else createOptions.db.withReadPreference(ReadPreference.valueOf(v, tags))
         val newCreateOptions = createOptions.copy(db = newDb)
         val newIterable = aggregate(options, newCreateOptions)
-        return AggregateIterableHelper(newIterable, evaluator, options, newCreateOptions)
+        return AggregateIterableHelper(newIterable, converter, options, newCreateOptions)
     }
 
     override fun maxTimeMS(v: Long) = set("maxTimeMS", v)
@@ -123,10 +125,10 @@ internal data class FindCreateOptions(val db: MongoDatabase,
                                       val find: Document)
 
 internal class FindIterableHelper(iterable: FindIterable<out Any?>,
-                                  evaluator: MongoShellEvaluator,
+                                  converter: MongoShellConverter,
                                   options: Document?,
                                   private val createOptions: FindCreateOptions?)
-    : BaseMongoIterableHelper<FindIterable<out Any?>>(iterable, evaluator, options) {
+    : BaseMongoIterableHelper<FindIterable<out Any?>>(iterable, converter, options) {
     override val converters = findConverters
     override val defaultConverter = findDefaultConverter
 
@@ -137,7 +139,7 @@ internal class FindIterableHelper(iterable: FindIterable<out Any?>,
         else createOptions.db.withReadPreference(ReadPreference.valueOf(v, tags))
         val newCreateOptions = createOptions.copy(db = newDb)
         val newIterable = find(options, newCreateOptions)
-        return FindIterableHelper(newIterable, evaluator, options, newCreateOptions)
+        return FindIterableHelper(newIterable, converter, options, newCreateOptions)
     }
 
     override fun allowPartialResults() = set("allowPartialResults", true)
@@ -163,10 +165,10 @@ internal class FindIterableHelper(iterable: FindIterable<out Any?>,
     }
 }
 
-internal fun helper(iterable: MongoIterable<out Any?>, evaluator: MongoShellEvaluator): BaseMongoIterableHelper<*> {
+internal fun helper(iterable: MongoIterable<out Any?>, converter: MongoShellConverter): BaseMongoIterableHelper<*> {
     return when (iterable) {
-        is FindIterable -> FindIterableHelper(iterable, evaluator, null, null)
-        is AggregateIterable -> AggregateIterableHelper(iterable, evaluator, null, null)
-        else -> MongoIterableHelper(iterable, evaluator, null)
+        is FindIterable -> FindIterableHelper(iterable, converter, null, null)
+        is AggregateIterable -> AggregateIterableHelper(iterable, converter, null, null)
+        else -> MongoIterableHelper(iterable, converter, null)
     }
 }
