+++ packages/java-shell/src/main/kotlin/com/mongodb/mongosh/service/Cursor.kt
@@ -1,13 +1,13 @@
 package com.mongodb.mongosh.service
 
 import com.mongodb.client.MongoCursor
-import com.mongodb.mongosh.MongoShellEvaluator
+import com.mongodb.mongosh.MongoShellConverter
 import com.mongodb.mongosh.result.DocumentResult
 import org.bson.Document
 import org.graalvm.polyglot.HostAccess
 import org.graalvm.polyglot.Value
 
-internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private val evaluator: MongoShellEvaluator) : ServiceProviderCursor {
+internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private val converter: MongoShellConverter) : ServiceProviderCursor {
     private var iterator: MongoCursor<out Any?>? = null
     private var closed = false
 
@@ -75,7 +75,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
         if (!v.hasMembers()) {
             throw IllegalArgumentException("Expected one argument of type object. Got: $v")
         }
-        val collation = toDocument(evaluator, v)
+        val collation = toDocument(converter, v)
         helper.collation(collation)
         return this
     }
@@ -96,7 +96,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
     @HostAccess.Export
     override fun explain(verbosity: String?): Any? {
         checkQueryNotExecuted()
-        return evaluator.toJs(helper.explain(verbosity))
+        return converter.toJs(helper.explain(verbosity))
     }
 
     @HostAccess.Export
@@ -105,7 +105,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
             throw IllegalArgumentException("Expected one argument of type function. Got: $func")
         }
         getOrCreateIterator().forEach { v ->
-            func.execute(evaluator.toJs(v))
+            func.execute(converter.toJs(v))
         }
     }
 
@@ -121,7 +121,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
         if (v.isString) {
             helper.hint(v.asString())
         } else if (v.hasMembers()) {
-            helper.hint(toDocument(evaluator, v))
+            helper.hint(toDocument(converter, v))
         }
         return this
     }
@@ -160,7 +160,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
         if (!v.hasMembers()) {
             throw IllegalArgumentException("Expected one argument of type object. Got: $v")
         }
-        helper.max(toDocument(evaluator, v))
+        helper.max(toDocument(converter, v))
         return this
     }
 
@@ -181,7 +181,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
         if (!v.hasMembers()) {
             throw IllegalArgumentException("Expected one argument of type object. Got: $v")
         }
-        helper.min(toDocument(evaluator, v))
+        helper.min(toDocument(converter, v))
         return this
     }
 
@@ -191,7 +191,7 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
     @HostAccess.Export
     override fun project(v: Value): ServiceProviderCursor {
         checkQueryNotExecuted()
-        helper.projection(toDocument(evaluator, v))
+        helper.projection(toDocument(converter, v))
         return this
     }
 
@@ -228,22 +228,22 @@ internal class Cursor(private var helper: BaseMongoIterableHelper<*>, private va
     @HostAccess.Export
     override fun sort(spec: Value): Cursor {
         checkQueryNotExecuted()
-        helper.sort(toDocument(evaluator, spec))
+        helper.sort(toDocument(converter, spec))
         return this
     }
 
     @HostAccess.Export
     override fun toArray(): Any? {
         checkQueryNotExecuted()
-        return evaluator.toJs(helper.toArray())
+        return converter.toJs(helper.toArray())
     }
 
     private fun checkQueryNotExecuted() {
         check(iterator == null) { "query already executed" }
     }
 
-    private fun toDocument(evaluator: MongoShellEvaluator, map: Value?): Document {
+    private fun toDocument(converter: MongoShellConverter, map: Value?): Document {
         return if (map == null || map.isNull) Document()
-        else (evaluator.extract(map) as DocumentResult).value
+        else (converter.toJava(map) as DocumentResult).value
     }
 }
